+++ lexforge/lexforge-mod/src/main/java/settingdust/lazyyyyy/mixin/forge/lazy_entity_renderers/cloudstorage/ClientProxyMixin.java
@@ -1,24 +1,25 @@
-package settingdust.lazyyyyy.mixin.forge.lazy_entity_renderers.taniwha;
+package settingdust.lazyyyyy.mixin.forge.lazy_entity_renderers.cloudstorage;
 
+import com.github.alexthe668.cloudstorage.CloudStorage;
+import com.github.alexthe668.cloudstorage.client.ClientProxy;
 import com.llamalad7.mixinextras.injector.v2.WrapWithCondition;
 import com.moulberry.mixinconstraints.annotations.IfModLoaded;
 import net.minecraft.client.renderer.entity.LivingEntityRenderer;
 import net.minecraft.client.renderer.entity.layers.RenderLayer;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import party.lemons.taniwha.forge.TForgeClientEvents;
 
-@IfModLoaded("taniwha")
-@Mixin(TForgeClientEvents.class)
-public class TForgeClientEventsMixin {
+@IfModLoaded(CloudStorage.MODID)
+@Mixin(ClientProxy.class)
+public class ClientProxyMixin {
     @WrapWithCondition(
-        method = "addLayers",
+        method = "onAddLayers",
         at = @At(
             value = "INVOKE",
             target = "Lnet/minecraft/client/renderer/entity/LivingEntityRenderer;addLayer(Lnet/minecraft/client/renderer/entity/layers/RenderLayer;)Z"
         )
     )
-    private static boolean lazyyyyy$nullCheck(final LivingEntityRenderer instance, final RenderLayer arg) {
+    private static boolean lazyyyyy$avoidNull(final LivingEntityRenderer instance, final RenderLayer arg) {
         return instance != null;
     }
 }